rename for tests to be in new tests folder
