+++ springdoc-openapi-webflux-ui/src/test/java/test/org/springdoc/ui/app32/SpringDocBehindProxyWithCustomUIPathTest.java
@@ -16,7 +16,7 @@
  *
  */
 
-package test.org.springdoc.ui.app31;
+package test.org.springdoc.ui.app32;
 
 import org.junit.jupiter.api.Test;
 import test.org.springdoc.ui.AbstractSpringDocTest;
@@ -27,9 +27,10 @@ import org.springframework.test.context.TestPropertySource;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @TestPropertySource(properties = {
-		"server.forward-headers-strategy=framework"
+		"server.forward-headers-strategy=framework",
+		"springdoc.swagger-ui.path=/foo/documentation/swagger.html"
 })
-public class SpringDocBehindProxyTest extends AbstractSpringDocTest {
+public class SpringDocBehindProxyWithCustomUIPathTest extends AbstractSpringDocTest {
 
 	private static final String X_FORWARD_PREFIX = "/path/prefix";
 
@@ -37,35 +38,27 @@ public class SpringDocBehindProxyTest extends AbstractSpringDocTest {
 	static class SpringDocTestApp {}
 
 	@Test
-	public void shouldServeSwaggerUIAtDefaultPath() {
-		webTestClient.get().uri("/webjars/swagger-ui/index.html").exchange()
-				.expectStatus().isOk();
+	public void shouldRedirectSwaggerUIFromCustomPath() {
+		webTestClient
+				.get().uri("/foo/documentation/swagger.html")
+				.header("X-Forwarded-Prefix", X_FORWARD_PREFIX)
+				.exchange()
+				.expectStatus().isFound()
+				.expectHeader().location("/path/prefix/foo/documentation/swagger-ui/index.html");
 	}
 
 	@Test
-	public void shouldReturnCorrectInitializerJS() throws Exception {
+	public void shouldReturnCorrectInitializerJS() {
 		webTestClient
-				.get().uri("/webjars/swagger-ui/swagger-initializer.js")
+				.get().uri("/foo/documentation/swagger-ui/swagger-initializer.js")
 				.header("X-Forwarded-Prefix", X_FORWARD_PREFIX)
 				.exchange()
 				.expectStatus().isOk()
 				.expectBody(String.class)
 				.consumeWith(response ->
 						assertThat(response.getResponseBody())
-								.contains("\"configUrl\" : \"/path/prefix/v3/api-docs/swagger-config\",")
+								.contains("\"configUrl\" : \\\"/path/prefix/v3/api-docs/swagger-config\\\",")
 				);
-	}
-
-	@Test
-	public void shouldCalculateOauthRedirectBehindProxy() throws Exception {
-		webTestClient
-				.get().uri("/v3/api-docs/swagger-config")
-				.header("X-Forwarded-Proto", "https")
-				.header("X-Forwarded-Host", "proxy-host")
-				.header("X-Forwarded-Prefix", X_FORWARD_PREFIX)
-				.exchange()
-				.expectStatus().isOk().expectBody()
-				.jsonPath("$.oauth2RedirectUrl").isEqualTo("https://proxy-host/path/prefix/swagger-ui/oauth2-redirect.html");
 	}
 
 	@Test
@@ -74,7 +67,8 @@ public class SpringDocBehindProxyTest extends AbstractSpringDocTest {
 				.get().uri("/v3/api-docs/swagger-config")
 				.header("X-Forwarded-Prefix", X_FORWARD_PREFIX)
 				.exchange()
-				.expectStatus().isOk().expectBody()
+				.expectStatus().isOk()
+				.expectBody()
 				.jsonPath("$.url")
 				.isEqualTo("/path/prefix/v3/api-docs")
 				.jsonPath("$.configUrl")
